requirements.txt
google-generativeai
Flask
Werkzeug
gunicorn # Productie WSGI server

# python-dotenv # Verwijderd, niet nodig op Render 